Move BuildConfig and dependencies into new module
